+++ screens/AccountScreens/Contact.js
@@ -1,4 +1,3 @@
-
 import React from 'react'; 
 import { Text, View, StyleSheet, Button} from 'react-native';
